Move AccessTokenHandler into own OAuthService
