+++ pkg/services/irma/irma_validator.go
@@ -5,7 +5,6 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"time"
 
 	"github.com/nuts-foundation/nuts-auth/pkg/services"
 
@@ -15,8 +14,6 @@ import (
 
 	core "github.com/nuts-foundation/nuts-go-core"
 
-	"github.com/google/uuid"
-
 	"github.com/dgrijalva/jwt-go"
 	nutscrypto "github.com/nuts-foundation/nuts-crypto/pkg"
 	cryptoTypes "github.com/nuts-foundation/nuts-crypto/pkg/types"
@@ -25,6 +22,9 @@ import (
 	irmaserver "github.com/privacybydesign/irmago/server"
 	"github.com/sirupsen/logrus"
 )
+
+var _ services.ContractSessionHandler = (*IrmaService)(nil)
+var _ services.ContractValidator = (*IrmaService)(nil)
 
 // IrmaService validates contracts using the irma logic.
 type IrmaService struct {
@@ -266,160 +266,6 @@ func convertPayloadToClaims(payload services.NutsIdentityToken) (map[string]inte
 // This is mainly a wrapper around the irma.IrmaSessionHandler.StartSession
 func (v IrmaService) StartSession(request interface{}, handler irmaserver.SessionHandler) (*irma.Qr, string, error) {
 	return v.IrmaSessionHandler.StartSession(request, handler)
-}
-
-// ParseAndValidateJwtBearerToken validates the jwt signature and returns the containing claims
-func (v IrmaService) ParseAndValidateJwtBearerToken(acString string) (*services.NutsJwtBearerToken, error) {
-	parser := &jwt.Parser{ValidMethods: []string{jwt.SigningMethodRS256.Name}}
-	token, err := parser.ParseWithClaims(acString, &services.NutsJwtBearerToken{}, func(token *jwt.Token) (i interface{}, e error) {
-		legalEntity, err := parseTokenIssuer(token.Claims.(*services.NutsJwtBearerToken).Issuer)
-		if err != nil {
-			return nil, err
-		}
-
-		// get public key
-		org, err := v.Registry.OrganizationById(legalEntity)
-		if err != nil {
-			return nil, err
-		}
-
-		pk, err := org.CurrentPublicKey()
-		if err != nil {
-			return nil, err
-		}
-
-		return pk.Materialize()
-	})
-
-	if token != nil && token.Valid {
-		if claims, ok := token.Claims.(*services.NutsJwtBearerToken); ok {
-			return claims, nil
-		}
-	}
-
-	return nil, err
-}
-
-// BuildAccessToken builds an access token based on the oauth claims and the identity of the user provided by the identityValidationResult
-// The token gets signed with the custodians private key and returned as a string.
-func (v IrmaService) BuildAccessToken(jwtBearerToken *services.NutsJwtBearerToken, identityValidationResult *services.ContractValidationResult) (string, error) {
-
-	if identityValidationResult.ValidationResult != services.Valid {
-		return "", fmt.Errorf("could not build accessToken: %w", errors.New("invalid contract"))
-	}
-
-	issuer := jwtBearerToken.Subject
-	if issuer == "" {
-		return "", fmt.Errorf("could not build accessToken: %w", errors.New("subject is missing"))
-	}
-
-	at := services.NutsAccessToken{
-		StandardClaims: jwt.StandardClaims{
-			// Expires in 15 minutes
-			ExpiresAt: time.Now().Add(time.Minute * 15).Unix(),
-			IssuedAt:  time.Now().Unix(),
-			Issuer:    issuer,
-			Subject:   jwtBearerToken.Issuer,
-		},
-		SubjectID: jwtBearerToken.SubjectID,
-		Scope:     jwtBearerToken.Scope,
-		// based on
-		// https://privacybydesign.foundation/attribute-index/en/pbdf.gemeente.personalData.html
-		// https://privacybydesign.foundation/attribute-index/en/pbdf.pbdf.email.html
-		// and
-		// https://openid.net/specs/openid-connect-basic-1_0.html#StandardClaims
-		FamilyName: identityValidationResult.DisclosedAttributes["gemeente.personalData.familyname"],
-		GivenName:  identityValidationResult.DisclosedAttributes["gemeente.personalData.firstnames"],
-		Prefix:     identityValidationResult.DisclosedAttributes["gemeente.personalData.prefix"],
-		Name:       identityValidationResult.DisclosedAttributes["gemeente.personalData.fullname"],
-		Email:      identityValidationResult.DisclosedAttributes["pbdf.email.email"],
-	}
-
-	var keyVals map[string]interface{}
-	inrec, _ := json.Marshal(at)
-	if err := json.Unmarshal(inrec, &keyVals); err != nil {
-		return "", err
-	}
-
-	// Sign with the private key of the issuer
-	token, err := v.Crypto.SignJWT(keyVals, cryptoTypes.KeyForEntity(cryptoTypes.LegalEntity{URI: issuer}))
-	if err != nil {
-		return token, fmt.Errorf("could not build accessToken: %w", err)
-	}
-
-	return token, err
-}
-
-// CreateJwtBearerToken creates a JwtBearerTokenResult containing a jwtBearerToken from a CreateJwtBearerTokenRequest.
-func (v IrmaService) CreateJwtBearerToken(request *services.CreateJwtBearerTokenRequest) (*services.JwtBearerTokenResult, error) {
-	jti, err := uuid.NewRandom()
-	if err != nil {
-		return nil, err
-	}
-
-	jwtBearerToken := services.NutsJwtBearerToken{
-		StandardClaims: jwt.StandardClaims{
-			//Audience:  endpoint.Identifier.String(),
-			ExpiresAt: time.Now().Add(30 * time.Minute).Unix(),
-			Id:        jti.String(),
-			IssuedAt:  time.Now().Unix(),
-			Issuer:    request.Actor,
-			NotBefore: 0,
-			Subject:   request.Custodian,
-		},
-		IdentityToken: request.IdentityToken,
-		SubjectID:     request.Subject,
-		Scope:         request.Scope,
-	}
-
-	var keyVals map[string]interface{}
-	inrec, _ := json.Marshal(jwtBearerToken)
-	if err := json.Unmarshal(inrec, &keyVals); err != nil {
-		return nil, err
-	}
-
-	signingString, err := v.Crypto.SignJWT(keyVals, cryptoTypes.KeyForEntity(cryptoTypes.LegalEntity{URI: request.Actor}))
-	if err != nil {
-		return nil, err
-	}
-
-	return &services.JwtBearerTokenResult{BearerToken: signingString}, nil
-}
-
-// ParseAndValidateAccessToken parses and validates a accesstoken string and returns a filled in NutsAccessToken.
-func (v IrmaService) ParseAndValidateAccessToken(accessToken string) (*services.NutsAccessToken, error) {
-	parser := &jwt.Parser{ValidMethods: []string{jwt.SigningMethodRS256.Name}}
-	token, err := parser.ParseWithClaims(accessToken, &services.NutsAccessToken{}, func(token *jwt.Token) (i interface{}, e error) {
-		legalEntity, err := parseTokenIssuer(token.Claims.(*services.NutsAccessToken).Issuer)
-		if err != nil {
-			return nil, err
-		}
-
-		// Check if the care provider which signed the token is managed by this node
-		if !v.Crypto.PrivateKeyExists(cryptoTypes.KeyForEntity(cryptoTypes.LegalEntity{URI: legalEntity.String()})) {
-			return nil, fmt.Errorf("invalid token: not signed by a care provider of this node")
-		}
-
-		// get public key
-		org, err := v.Registry.OrganizationById(legalEntity)
-		if err != nil {
-			return nil, err
-		}
-
-		pk, err := org.CurrentPublicKey()
-		if err != nil {
-			return nil, err
-		}
-
-		return pk.Materialize()
-	})
-
-	if token != nil && token.Valid {
-		if claims, ok := token.Claims.(*services.NutsAccessToken); ok {
-			return claims, nil
-		}
-	}
-	return nil, err
 }
 
 func parseTokenIssuer(issuer string) (core.PartyID, error) {
